sombreando/backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-otp==1.2.2
qrcode==7.4.2
cryptography==41.0.7

# Email
django-ses==3.5.0

# Payment Integration
mercadopago==2.2.1

# API Documentation
drf-spectacular==0.26.5

# Utilities
python-decouple==3.8
Pillow==10.0.1
celery==5.3.4
redis==5.0.1

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
black==23.10.1
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

